[refactor] move recommender to its own sub module
