+++ server/db/config.js
@@ -1,12 +1,10 @@
-require('dotenv').config();
+require('dotenv').config(); //use for development
 var Sequelize = require('sequelize');
 var User = require('./models/users.js');
-// var Skill = require('./models/skills.js');
 var Message = require('./models/messages.js');
-// var UserSkillJoin = require('./models/userSkills.js');
 var Invitation = require('./models/invitations.js');
 
-var sequelize = new Sequelize(process.env.DATABASE_URL); //Uncomment for real deployment
+var sequelize = new Sequelize(process.env.DATABASE_URL);
 
 sequelize
   .authenticate()
