+++ server/db/helpers.js
@@ -1,4 +1,5 @@
 var db = require('./config.js');
+var rec = require('../recommender/recommend');
 var helpers = {};
 
 //---------------AUTHENTICATION----------------------
@@ -72,43 +73,15 @@ helpers.updateUser = function(profileUpdateObj) {
   });
 };
 
-
 //------------------GET RECOMMENDED USERS-------------------------
 
 helpers.getAllUsers = function(username) {
   return db.User.findAll()
   .then(function(usersArray) {
-    var result = helpers.getRecommendations(usersArray, username);
+    var result = rec.getRecommendations(usersArray, username);
     // console.log('HERE ARE ALL USERS', usersArray);
     return result;
   });
-};
-
-//more functions need to be written here for the recommender
-
-helpers.getRecommendations = function(usersArray, username) {
-  //get the object with the current user and take it out
-  //of the usersArray
-  var current_user = {};
-  for (var i = 0; i < usersArray.length; i++) {
-    if (usersArray[i].dataValues.username === username) {
-      current_user = usersArray[i].dataValues;
-      usersArray.splice(i, 1);
-    }
-  }
-
-  usersArray.sort(function(a, b){
-    return b.dataValues.karmaPoints - a.dataValues.karmaPoints;
-  });
-  usersArray.forEach(function(item){
-    console.log(item.dataValues.karmaPoints, item.dataValues.username);
-  });
-
-  // console.log(usersArray);
-  //we want to sort usersArray by comparing every elt to current_user and
-  //assigning some sort of ranking
-
-  return usersArray;
 };
 
 module.exports = helpers;
